refactor code, improve performance, fix bugs
Ok, this commit change a lot of this program

First of all, I needed to refactor the code to support
suspending/unsuspending of cards that were not added today (ie. cards
that are not "added:1")

I also made some structural changes to leverage batch editing of
cards (this improves performance a BUNCH)

And, last but not least, I fixed some bugs when suspending/unsuspending cards.